refactor: move opcode handlers to appropriate modules.
