+++ actors/evm/src/interpreter/execution.rs
@@ -88,11 +88,11 @@ pub fn execute<'r, BS: Blockstore, RT: Runtime<BS>>(
             OpCode::CALLDATACOPY => call::calldatacopy(runtime)?,
             OpCode::CODESIZE => call::codesize(&mut runtime.stack, bytecode.as_ref()),
             OpCode::CODECOPY => call::codecopy(runtime, bytecode.as_ref())?,
-            OpCode::EXTCODESIZE => storage::extcodesize(runtime, system)?,
-            OpCode::EXTCODECOPY => memory::extcodecopy(runtime, system)?,
+            OpCode::EXTCODESIZE => ext::extcodesize(runtime, system)?,
+            OpCode::EXTCODECOPY => ext::extcodecopy(runtime, system)?,
             OpCode::RETURNDATASIZE => control::returndatasize(runtime),
             OpCode::RETURNDATACOPY => control::returndatacopy(runtime)?,
-            OpCode::EXTCODEHASH => storage::extcodehash(runtime, system)?,
+            OpCode::EXTCODEHASH => ext::extcodehash(runtime, system)?,
             OpCode::BLOCKHASH => context::blockhash(runtime, system)?,
             OpCode::ORIGIN => context::origin(runtime, system),
             OpCode::COINBASE => context::coinbase(runtime, system)?,
@@ -100,10 +100,10 @@ pub fn execute<'r, BS: Blockstore, RT: Runtime<BS>>(
             OpCode::TIMESTAMP => context::timestamp(runtime, system)?,
             OpCode::NUMBER => context::block_number(runtime, system)?,
             OpCode::DIFFICULTY => context::difficulty(runtime, system)?,
-            OpCode::GASLIMIT => context::gas_limit(runtime, system)?,
+            OpCode::GASLIMIT => context::gas_limit(runtime, system),
             OpCode::CHAINID => context::chain_id(runtime, system)?,
-            OpCode::BASEFEE => context::base_fee(runtime, system)?,
-            OpCode::SELFBALANCE => storage::selfbalance(runtime, system)?,
+            OpCode::BASEFEE => context::base_fee(runtime, system),
+            OpCode::SELFBALANCE => state::selfbalance(runtime, system)?,
             OpCode::POP => stack::pop(&mut runtime.stack),
             OpCode::MLOAD => memory::mload(runtime)?,
             OpCode::MSTORE => memory::mstore(runtime)?,
