+++ actors/evm/src/interpreter/instructions/mod.rs
@@ -4,8 +4,11 @@ pub mod boolean;
 pub mod call;
 pub mod context;
 pub mod control;
+pub mod ext;
 pub mod hash;
+pub mod lifecycle;
 pub mod log;
 pub mod memory;
 pub mod stack;
+pub mod state;
 pub mod storage;
