+++ actors/evm/src/interpreter/instructions/state.rs
@@ -1,0 +1,21 @@
+use {
+    crate::interpreter::{ExecutionState, StatusCode, System, U256},
+    fil_actors_runtime::runtime::Runtime,
+    fvm_ipld_blockstore::Blockstore,
+};
+
+#[inline]
+pub fn balance<'r, BS: Blockstore, RT: Runtime<BS>>(
+    _state: &mut ExecutionState,
+    _platform: &'r System<'r, BS, RT>,
+) -> Result<(), StatusCode> {
+    todo!()
+}
+
+#[inline]
+pub fn selfbalance<'r, BS: Blockstore, RT: Runtime<BS>>(
+    _state: &mut ExecutionState,
+    _platform: &'r System<'r, BS, RT>,
+) -> Result<(), StatusCode> {
+    todo!()
+}
