+++ actors/evm/src/interpreter/instructions/storage.rs
@@ -33,51 +33,3 @@ pub fn sstore<'r, BS: Blockstore, RT: Runtime<BS>>(
     platform.set_storage(location, opt_value)?;
     Ok(())
 }
-
-#[inline]
-pub fn balance<'r, BS: Blockstore, RT: Runtime<BS>>(
-    _state: &mut ExecutionState,
-    _platform: &'r System<'r, BS, RT>,
-) -> Result<(), StatusCode> {
-    todo!()
-}
-
-#[inline]
-pub fn selfbalance<'r, BS: Blockstore, RT: Runtime<BS>>(
-    _state: &mut ExecutionState,
-    _platform: &'r System<'r, BS, RT>,
-) -> Result<(), StatusCode> {
-    todo!()
-}
-
-#[inline]
-pub fn extcodesize<'r, BS: Blockstore, RT: Runtime<BS>>(
-    _state: &mut ExecutionState,
-    _platform: &'r System<'r, BS, RT>,
-) -> Result<(), StatusCode> {
-    todo!()
-}
-
-pub fn extcodehash<'r, BS: Blockstore, RT: Runtime<BS>>(
-    _state: &mut ExecutionState,
-    _platform: &'r System<'r, BS, RT>,
-) -> Result<(), StatusCode> {
-    todo!();
-}
-
-#[inline]
-pub fn create<'r, BS: Blockstore, RT: Runtime<BS>>(
-    _state: &mut ExecutionState,
-    _platform: &'r System<'r, BS, RT>,
-    _create2: bool,
-) -> Result<(), StatusCode> {
-    todo!()
-}
-
-#[inline]
-pub fn selfdestruct<'r, BS: Blockstore, RT: Runtime<BS>>(
-    _state: &mut ExecutionState,
-    _platform: &'r System<'r, BS, RT>,
-) -> Result<(), StatusCode> {
-    todo!()
-}
